requirements.txt
aioredis==1.3.1
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.12.3
Automat==20.2.0
certifi==2020.12.5
cffi==1.14.4
channels==3.0.3
channels-redis==3.2.0
chardet==4.0.0
click==7.1.2
constantly==15.1.0
cryptography==3.3.1
daphne==3.0.1
Django==3.1.5
django-cors-headers==3.7.0
djangorestframework==3.12.2
djutils==1.0.13
Flask==1.1.2
hiredis==1.1.0
hyperlink==21.0.0
idna==2.10
incremental==17.5.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
msgpack==1.0.2
msgpack-python==0.5.6
Pillow==8.1.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==20.0.1
pytz==2020.5
redis==2.10.6
requests==2.25.1
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
Twisted
txaio==20.12.1
urllib3==1.26.3
websocket-client==0.57.0
Werkzeug==1.0.1
zope.interface==5.2.0
